+++ resources/views/base/base.blade.php
@@ -180,8 +180,6 @@
             break;
     }
     @endif
-
-    @include('sweetalert::alert')
 </script>
 
 <!--
@@ -200,8 +198,6 @@ return false;
 </script> -->
 {{--end::TOASTR options --}}
 <livewire:scripts />
-<script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
-<x-livewire-alert::scripts />
 <script src="{{ asset('vendor/livewire-alert/livewire-alert.js') }}"></script>
 <x-livewire-alert::flash />
 </body>