+++ resources/views/emails/tractors.blade.php
@@ -1,0 +1,8 @@
+@component('mail::message')
+# Tractor Added
+@component('mail::panel')
+Tractor {{ $request->tractor_id }} has been added to the system.
+@endcomponent
+Thanks,
+Unknown TMS
+@endcomponent